🔨refactor: Airframe の初期化を AppGuardian で行う
